requirement.txt
langchain_openai
langchain
streamlit
faiss-cpu
python-dotenv
watchdog
langchain_community
unstructured